Update to 2.0.3.0
．One vertex can be now control by more than 4 bones, but this kind of mesh will use software skinning only.
．Remove every GL13.glClientActiveTexture(COLOR_MAP_INDEX); for being unnecessary.
．Fix rendering glitch and game crash on GL20 profile.
