+++ src/main/java/com/modularmods/mcgltf/RenderedGltfSceneGL20.java
@@ -1,12 +1,18 @@
 package com.modularmods.mcgltf;
 
+import org.lwjgl.opengl.GL11;
 import org.lwjgl.opengl.GL15;
+import org.lwjgl.opengl.GL20;
 
 public class RenderedGltfSceneGL20 extends RenderedGltfScene {
 
 	@Override
 	public void renderForVanilla() {
 		vanillaRenderCommands.forEach(Runnable::run);
+		
+		GL11.glDisableClientState(GL11.GL_VERTEX_ARRAY);
+		GL11.glDisableClientState(GL11.GL_NORMAL_ARRAY);
+		GL20.glDisableVertexAttribArray(RenderedGltfModel.at_tangent);
 		
 		GL15.glBindBuffer(GL15.GL_ARRAY_BUFFER, 0);
 		GL15.glBindBuffer(GL15.GL_ELEMENT_ARRAY_BUFFER, 0);
@@ -16,6 +22,10 @@ public class RenderedGltfSceneGL20 extends RenderedGltfScene {
 	@Override
 	public void renderForShaderMod() {
 		shaderModRenderCommands.forEach(Runnable::run);
+		
+		GL11.glDisableClientState(GL11.GL_VERTEX_ARRAY);
+		GL11.glDisableClientState(GL11.GL_NORMAL_ARRAY);
+		GL20.glDisableVertexAttribArray(RenderedGltfModel.at_tangent);
 		
 		GL15.glBindBuffer(GL15.GL_ARRAY_BUFFER, 0);
 		GL15.glBindBuffer(GL15.GL_ELEMENT_ARRAY_BUFFER, 0);
